Change the effect to use the deployment event subscription
Caveat: This requires the updated version of deployment.lua instead of the present one (at time of submission), due to the oversight which causes it to only trigger once the first Mech (Pawn Id 0) has landed.

There is also a delay created by the skill-effect causing the Board to be in a busy state.

The only solution to this (to the best of my knowledge) would be to create a global counter that increments until all three Mechs have deployed, upon which then you choose to trigger the effect without creating a weird delay between Mechs landing. If you want this solution to be implemented, let me know and I will send it through.